move residents controller
