--- app/views/exchanges/residents/_remote_search_form.html.haml
@@ -1,2 +1,0 @@
-= render 'shared/error_messages', :object => @person
-= render "exchanges/residents/resident_search_form", f: f
