--- app/views/exchanges/residents/_resident_search_form.html.haml
@@ -1,4 +1,0 @@
-= render 'shared/resident_search_fields', f: f
-
-%br
-#person_search_results
