+++ app/views/individual_market/exchanges/residents/_remote_search_form.html.haml
@@ -1,0 +1,2 @@
+= render 'shared/error_messages', :object => @person
+= render "individual_market/exchanges/residents/resident_search_form", f: f
