+++ app/views/individual_market/exchanges/residents/_resident_application_link.html.erb
@@ -1,6 +1,6 @@
 <%=pundit_span Family, :updateable? %>
      <ul class='col-xs-12 col-md-5 col-lg-5'>
-       <li> <%=link_to 'New CoverAll DC Application', begin_resident_enrollment_exchanges_residents_path(
+       <li> <%=link_to 'New CoverAll DC Application', begin_resident_individual_market_enrollment_exchanges_residents_path(
        original_application_type: 'paper')  %> </li>
      </ul>
 </span>
