+++ app/views/individual_market/exchanges/residents/_resident_search_form.html.haml
@@ -1,0 +1,4 @@
+= render 'shared/individual_market/resident_search_fields', f: f
+
+%br
+#person_search_results
