+++ app/views/individual_market/exchanges/residents/index.html.erb
