+++ app/views/individual_market/exchanges/residents/index.js.erb
