+++ app/views/individual_market/exchanges/residents/match.html.haml
@@ -4,9 +4,9 @@
   #new_person_wrapper
     .row
       .col-lg-7.col-md-7.col-sm-8.col-xs-12
-        = form_for @person, { :as => :person, :url => match_exchanges_residents_path, :method => :post } do |f|
+        = form_for @person, { :as => :person, :url => match_individual_market_exchanges_residents_path, :method => :post } do |f|
           %h3 Personal Information
-          = render 'exchanges/residents/remote_search_form', f: f
+          = render 'individual_market/exchanges/residents/remote_search_form', f: f
           = render "no_match"
       .col-lg-3.col-md-3.col-sm-3.col-xs-12.col-md-offset-2.col-sm-offset-1.right-section
-        = render  partial: 'shared/resident_progress', locals: {step: '1', search_page: true}
+        = render  partial: 'shared/individual_market/resident_progress', locals: {step: '1', search_page: true}
