+++ app/views/individual_market/exchanges/residents/search.html.haml
@@ -3,11 +3,11 @@
 .container.pd-container
   #new_person_wrapper
     .row
-      = form_for @person, { :as => :person, :url => match_exchanges_residents_path, :method => :post } do |f|
+      = form_for @person, { :as => :person, :url => match_individual_market_exchanges_residents_path, :method => :post } do |f|
         .col-lg-7.col-md-7.col-sm-8.col-xs-12
           %h1 Personal Information
           %h4 First, we need to verify whether you're new to #{site_short_name} or affiliated with another account. Please enter your personal information. When you're finished, select CONTINUE.
           %p.memo * = required field
-          = render 'exchanges/residents/remote_search_form', f: f
+          = render 'individual_market/exchanges/residents/remote_search_form', f: f
         .col-lg-3.col-md-3.col-sm-3.col-xs-12.col-md-offset-2.col-sm-offset-1.right-section
-          = render  partial: 'shared/resident_progress', locals: {step: '1'}
+          = render  partial: 'shared/individual_market/resident_progress', locals: {step: '1'}
