+++ app/views/shared/individual_market/_resident_progress.html.haml
